requirements.txt
Django==4.1.1
psycopg2-binary==2.9.3
celery==5.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
PyJWT==2.5.0
pytz==2022.2.1
invoke==1.7.1
redis==4.3.4
beautifulsoup4==4.11.1
requests==2.28.1
fake-useragent==0.1.11
